+++ cfg/method_name.go
@@ -4,4 +4,5 @@ const (
 	getAvailableConfigs   = "/api/config/module/get_modules_info"
 	getConfigByModuleName = "/api/config/config/get_config_by_instance_uuid_and_module_name"
 	createUpdateConfig    = "/api/config/config/create_update_config"
+	getSchemaByModuleId   = "/api/config/schema/get_by_module_id"
 )