+++ flag/common.go
@@ -3,7 +3,8 @@ package flag
 import "github.com/codegangsta/cli"
 
 var (
-	Host  = cli.StringFlag{Name: GateHostName}
-	Uuid  = cli.StringFlag{Name: InstanceUuidName}
-	Color = cli.BoolFlag{Name: ColorName}
+	Host   = cli.StringFlag{Name: GateHostName, Usage: "overrides gateHost"}
+	Uuid   = cli.StringFlag{Name: InstanceUuidName, Usage: "overrides instanceUuid"}
+	Color  = cli.BoolFlag{Name: ColorName, Usage: "colorize the json for outputing to screen"}
+	Unsafe = cli.BoolFlag{Name: UnsafeName, Usage: "disable checking schema for set configuration"}
 )