+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## v1.1.0
+* fix status response
+* add checking schema to set configuration
+* add unsafe flag
+* update readme
 ## v1.0.1
 * fix spaces in strings